requirements.txt
discord-py
pillow
PyNacl
requests
html2text
